+++ src/buckstore.manager.service.domain/Aggregates/ProductAggregate/Product.cs
@@ -8,16 +8,16 @@ namespace buckstore.manager.service.domain.Aggregates.ProductAggregate
         public string Name => _name;
         private string _description;
         public string Description => _description;
-        private double _price;
-        public double Price => _price;
+        private decimal _price;
+        public decimal Price => _price;
         private int _stockQuantity;
         public int Stock => _stockQuantity;
         private int _categoryId;
         public ProductCategory Category { get; private set; }
-        
+
         protected Product() { }
 
-        public Product(string name, string description, double price, int stock, int categoryId)
+        public Product(string name, string description, decimal price, int stock, int categoryId)
         {
             _name = name;
             _description = description;
@@ -41,12 +41,12 @@ namespace buckstore.manager.service.domain.Aggregates.ProductAggregate
             return _stockQuantity > 0;
         }
 
-        public void ChangeProductPrice(double newPrice)
+        public void ChangeProductPrice(decimal newPrice)
         {
             _price = newPrice;
         }
 
-        public void UpdateProduct(string name, string description, double price, int stock, int categoryId)
+        public void UpdateProduct(string name, string description, decimal price, int stock, int categoryId)
         {
             _name = name;
             _description = description;
